Voting representation done. Needs html and css fixing.
If you click on a voting process you can see the proposalsets
with proposals that are inside listed with truncated description
and current votes count, that you can click to vote on. If a proposalset
is linked to a debate, a link to that debate will be shown too. Also, side by
side are the unlinked proposals, that also have description and current
votes count and you can click on the proposal so check it out and vote.