LocationSelector2: Reset hidden fields when address fields are changed
This fixes the problem with "Invalid Location" when first selecting all
Lx and then filling in the address fields - which wrongly retained the ID of the
lowest selected Lx as ID for the specific location instead of setting it
to "dummy" in order to create a new location (Tested that it does not overwrite
an existing specific location ID when filling the address fields).

This may need some additional logic to prune orphaned specific locations as you
are now able to reset to the lowest selected Lx by clearing the address fields
(which would currently leave the specific location in the DB without
anything referencing it).

Also tested that it returns to the specific location when you first clear,
but then fill the address fields again (rather than creating a new
location), so it doesn't create orphaned locations just because of the
order in which the user fills in the fields.