Display error in map component if there is no GIS config
Displaying the error in the map component allows the rest of the webpage
to load normally and removes an unsafe redirect from a lower level
function, as suggested in #1513.

This commit also moves the instantiation of the loader component into
the `_setup()` method after we've fetched the config, to ensure that the
error message doesn't appear with a permanent loading icon above it.